[KYUUBI #3934] Merge the test class to TrinoContextSuite
