+++ kyuubi-server/src/test/scala/org/apache/kyuubi/server/trino/api/TrinoContextSuite.scala
@@ -17,13 +17,23 @@
 
 package org.apache.kyuubi.server.trino.api
 
+import java.net.URI
 import java.time.ZoneId
+import javax.ws.rs.core.MediaType
+
+import scala.collection.JavaConverters._
 
 import io.trino.client.ProtocolHeaders.TRINO_HEADERS
+import org.apache.hive.service.rpc.thrift.TProtocolVersion.HIVE_CLI_SERVICE_PROTOCOL_V9
+import org.scalatest.concurrent.PatienceConfiguration.Timeout
+import org.scalatest.time.SpanSugar.convertIntToGrainOfTime
 
-import org.apache.kyuubi.KyuubiFunSuite
+import org.apache.kyuubi.{KyuubiFunSuite, RestFrontendTestHelper}
+import org.apache.kyuubi.events.KyuubiOperationEvent
+import org.apache.kyuubi.operation.{FetchOrientation, OperationHandle}
+import org.apache.kyuubi.operation.OperationState.{FINISHED, OperationState}
 
-class TrinoContextSuite extends KyuubiFunSuite {
+class TrinoContextSuite extends KyuubiFunSuite with RestFrontendTestHelper {
   import TrinoContext._
 
   test("create trino request context with header") {
@@ -67,4 +77,78 @@ class TrinoContextSuite extends KyuubiFunSuite {
     assert(actual == expectedTrinoContext)
   }
 
+  test("test convert") {
+    val opHandle = getOpHandle("select 1")
+    val opHandleStr = opHandle.identifier.toString
+    checkOpState(opHandleStr, FINISHED)
+
+    val metadataResp = fe.be.getResultSetMetadata(opHandle)
+    val tRowSet = fe.be.fetchResults(opHandle, FetchOrientation.FETCH_NEXT, 1000, false)
+    val status = fe.be.getOperationStatus(opHandle)
+
+    val uri = new URI("sfdsfsdfdsf")
+    val results = TrinoContext
+      .createQueryResults("/xdfd/xdf", uri, uri, status, Option(metadataResp), Option(tRowSet))
+
+    print(results.toString)
+    assert(results.getColumns.get(0).getType.equals("INT_TYPE"))
+    assert(results.getData.asScala.last.get(0) == 1)
+  }
+
+  test("test convert from table") {
+    initSql("CREATE DATABASE IF NOT EXISTS INIT_DB")
+    initSql("CREATE TABLE IF NOT EXISTS INIT_DB.test(a int) USING CSV;")
+    initSql("INSERT INTO INIT_DB.test VALUES (2)")
+
+    val opHandle = getOpHandle("SELECT * FROM INIT_DB.test")
+    val opHandleStr = opHandle.identifier.toString
+    checkOpState(opHandleStr, FINISHED)
+
+    val metadataResp = fe.be.getResultSetMetadata(opHandle)
+    val tRowSet = fe.be.fetchResults(opHandle, FetchOrientation.FETCH_NEXT, 1000, false)
+    val status = fe.be.getOperationStatus(opHandle)
+
+    val uri = new URI("sfdsfsdfdsf")
+    val results = TrinoContext
+      .createQueryResults("/xdfd/xdf", uri, uri, status, Option(metadataResp), Option(tRowSet))
+
+    print(results.toString)
+    assert(results.getColumns.get(0).getType.equals("INT_TYPE"))
+    assert(results.getData.asScala.last.get(0) == 2)
+  }
+
+  def getOpHandleStr(statement: String = "show tables"): String = {
+    getOpHandle(statement).identifier.toString
+  }
+
+  def getOpHandle(statement: String = "show tables"): OperationHandle = {
+    val sessionHandle = fe.be.openSession(
+      HIVE_CLI_SERVICE_PROTOCOL_V9,
+      "admin",
+      "123456",
+      "localhost",
+      Map("testConfig" -> "testValue"))
+
+    if (statement.nonEmpty) {
+      fe.be.executeStatement(sessionHandle, statement, Map.empty, runAsync = false, 5000)
+    } else {
+      fe.be.getCatalogs(sessionHandle)
+    }
+  }
+
+  private def checkOpState(opHandleStr: String, state: OperationState): Unit = {
+    eventually(Timeout(5.seconds)) {
+      val response = webTarget.path(s"api/v1/operations/$opHandleStr/event")
+        .request(MediaType.APPLICATION_JSON_TYPE).get()
+      assert(response.getStatus === 200)
+      val operationEvent = response.readEntity(classOf[KyuubiOperationEvent])
+      assert(operationEvent.state === state.name())
+    }
+  }
+
+  private def initSql(sql: String): Unit = {
+    val initOpHandle = getOpHandle(sql)
+    val initOpHandleStr = initOpHandle.identifier.toString
+    checkOpState(initOpHandleStr, FINISHED)
+  }
 }
